Move ConfirmedCasesNumberCard to components/index/cards
