+++ components/index/cards/ConfirmedCasesNumberCard.vue
